Split Auth.js into Login.js and Register.js
